requirements.txt
# FastAPI Backend (Lightweight)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# ML Inference (CPU-Only for Railway)
onnxruntime>=1.16.0
numpy>=1.24.0
pillow>=10.0.0

# Data Validation & Serialization
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6

# Security & Rate Limiting
slowapi>=0.1.9

# Optional: Redis support for production rate limiting
redis>=5.0.0

# Logging & Monitoring
python-json-logger>=2.0.7

# Environment management
python-dotenv>=1.0.0